ADD: Get, Patch and Delete task

+++ src/task/task.controller.ts
@@ -1,34 +1,50 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Query, HttpException, HttpStatus } from '@nestjs/common';
 import { TaskService } from './task.service';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
+import { PaginateDto } from './dto/paginate.dto';
 
 @Controller('tasks')
 export class TaskController {
   constructor(private readonly taskService: TaskService) {}
 
   @Post()
-  create(@Body() createTaskDto: CreateTaskDto) {
-    return this.taskService.create(createTaskDto);
+  async create(@Body() createTaskDto: CreateTaskDto) {
+    return await this.taskService.create(createTaskDto);
   }
 
   @Get()
-  findAll() {
-    return this.taskService.findAll();
+  async findAll(@Query() paginate: PaginateDto) {
+    return await this.taskService.findAll(paginate);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.taskService.findOne(+id);
+  async findOne(@Param('id') id: string) {
+    const task = await this.taskService.findOne(+id);
+
+    if (!task)
+      throw new HttpException('Task not found', HttpStatus.NOT_FOUND);
+
+    return task;
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateTaskDto: UpdateTaskDto) {
-    return this.taskService.update(+id, updateTaskDto);
+  async update(@Param('id') id: string, @Body() updateTaskDto: UpdateTaskDto) {
+    const task = await this.taskService.findOne(+id);
+
+    if (!task)
+      throw new HttpException('Task not found', HttpStatus.NOT_FOUND);
+
+    return await this.taskService.update(+id, updateTaskDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.taskService.remove(+id);
+  async remove(@Param('id') id: string) {
+    const task = await this.taskService.findOne(+id);
+
+    if (!task)
+      throw new HttpException('Task not found', HttpStatus.NOT_FOUND);
+    
+    return await this.taskService.remove(+id);
   }
 }
